ats_python_project/requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# NLP and text processing
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Data handling
pandas==2.1.4
numpy==1.25.2
sqlalchemy==2.0.23
aiosqlite==0.19.0

# LM Studio integration
openai==1.3.7
httpx==0.25.2

# Utilities
click==8.1.7
rich==13.7.0
loguru==0.7.2
pyyaml==6.0.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
